+++ src/main/kotlin/race/domain/FixedMoveCondition.kt
@@ -1,4 +1,4 @@
-package race.model
+package race.domain
 
 class FixedMoveCondition(private val canGoStraight: Boolean) : MoveCondition {
     override fun canGoStraight(): Boolean {
