+++ src/main/kotlin/race/domain/RandomMoveCondition.kt
@@ -1,4 +1,4 @@
-package race.model
+package race.domain
 
 import kotlin.random.Random
